pyproject.toml
[tool.poetry]
name = "myenv-iib"
version = "0.1.0"
description = "Poetry Environment for IIB Masters Project"
authors = ["ehp24 <ep590@cantab.ac.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
scipy = "^1.12.0"
glob2 = "^0.7"
laspy = "^2.5.3"
pillow = "^10.2.0"
pytest = "^8.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
